extract components from Devices
