+++ messages/en.json
@@ -15,8 +15,26 @@
   "screen.Home.title": {
     "message": "Mapeo Mobile Testing"
   },
+  "screen.sync.Devices.localDeviceDescription": {
+    "message": "These devices are on your project and on the same wifi network as you."
+  },
+  "screen.sync.Devices.localDevices": {
+    "message": "Local Devices"
+  },
   "screen.sync.Devices.searching": {
     "message": "Searching for devices"
+  },
+  "screen.sync.DevicesList.deviceName": {
+    "message": "Device Name"
+  },
+  "screen.sync.DevicesList.lastSynced": {
+    "message": "Last Synced"
+  },
+  "screen.sync.DevicesList.localDevices": {
+    "message": "Local Devices"
+  },
+  "screen.sync.DevicesList.sync": {
+    "message": "Sync"
   },
   "screen.sync.ProjectInfo.connectedNoNetworkName": {
     "message": "Connected (no network name)"
@@ -26,21 +44,6 @@
   },
   "screen.sync.index.navTitle": {
     "message": "Synchronize"
-  },
-  "screen.sync.main.deviceName": {
-    "message": "Device Name"
-  },
-  "screen.sync.main.lastSynced": {
-    "message": "Last Synced"
-  },
-  "screen.sync.main.localDeviceDescription": {
-    "message": "These devices are on your project and on the same wifi network as you."
-  },
-  "screen.sync.main.localDevices": {
-    "message": "Local Devices"
-  },
-  "screen.sync.main.sync": {
-    "message": "Sync"
   },
   "screens.sync.DeviceSyncInfo.close": {
     "message": "Close"
